Added a log in controller and fixed minor issues

+++ movies-omdb-rest/src/main/java/gr/aueb/cf/movies/rest/UserMovieController.java
@@ -7,7 +7,10 @@ import gr.aueb.cf.movies.model.User;
 import gr.aueb.cf.movies.service.util.JPAHelper;
 import jakarta.persistence.EntityManager;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
@@ -15,12 +18,19 @@ import javax.ws.rs.core.Response;
 public class UserMovieController {
 
     @POST
-    @Path("/{userId}/movies")
+    @Path("/{id}/movies")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response addMovieToWatchlist(@PathParam("userId") Long userId, Movie movie) {
+    public Response addMovieToWatchlist(@PathParam("id") Long userId, Movie movie, @Context HttpServletRequest request) {
         try {
             IUserDAO userDAO = new UserDAOImpl();
+            User user = userDAO.getById(userId);
+            HttpSession session = request.getSession(false);
+
+            if (session == null || session.getAttribute("userId") == null || !session.getAttribute("userId").equals(userId)) {
+                return Response.status(Response.Status.UNAUTHORIZED).build();
+            }
+
             userDAO.addMovie(userId, movie);
             return Response.status(Response.Status.CREATED).entity(movie).build();
         } catch (Exception e) {
@@ -29,12 +39,24 @@ public class UserMovieController {
     }
 
     @DELETE
-    @Path("/{userId}/movies")
+    @Path("/{id}/movies")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response removeMovieFromUserList(@PathParam("userId") Long userId, Movie movie) {
-        IUserDAO userDAO = new UserDAOImpl();
-        userDAO.removeMovie(userId, movie);
-        return Response.ok().entity(userDAO.getById(userId)).build();
+    public Response removeMovieFromWatchlist(@PathParam("id") Long userId, Movie movie, @Context HttpServletRequest request) {
+        try {
+            IUserDAO userDAO = new UserDAOImpl();
+            User user = userDAO.getById(userId);
+            HttpSession session = request.getSession(false);
+
+            if (session == null || session.getAttribute("userId") == null || !session.getAttribute("userId").equals(userId)) {
+                return Response.status(Response.Status.UNAUTHORIZED).build();
+            }
+
+            userDAO.removeMovie(userId, movie);
+            return Response.ok().entity(userDAO.getById(userId)).build();
+        } catch (Exception e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        }
     }
+
 }
